gc_backend/requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==10.2.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2==2.9.9
faker==16.0.0
pandas
numpy
asyncio
mistralai